The great Thrust index type fix: workaround for set algo tests on windows.
There's a particular configuration where the ifdef'd out part of the
test fails on our internal CI with a driver fault; waive this test on
windows to unblock landing the bulk of the changes.

Also add two missing includes to the adjacent_diff test. Apparently it
compiles on AMD64 Linux, but not ARM Linux and not on AMD64 Windows? We
should take a look at that one day.


[ROCm/rocThrust commit: 97689e9c5b8cf6d02e4662be0ef3a8880a773dd0]