Address cppcheck warnings - 1/2 (#342)
* Remove unused variables in cblas_labrd

* Initialize benchmark arguments

There wasn't actually anything wrong with the original code, but
cppcheck was complaining. The initialization at the declaration is
redundant, but it doesn't hurt.

* Add const to rocblas_nan_rng member functions

* Reduce variable scope

For primitive types, it doesn't matter whether you declare them in the
loop or outside. LLVM has an optimization pass for this transformation,
so the resulting machine code should be identical in release builds.

* Comment out unused variables in gesvd

[ROCm/rocSOLVER commit: 1c6be9fca625b576d8d806829defbbd5424181d8]