Do not specialize half to avoid compilation failures with it (#345)
* Add missing includes in device algorithm headers

Add any obviously missing standard library or rocprim headers to
the device headers.
The added headers are based clangd's error messages when opening only
the headers, and from a very quick glance to find others that are only
transitively included, but the symbol's are used in the file.

* Change `sample_to_bin_even` half specialization

- Change `std::is_floating_point` to `rocprim::is_floating_point` to
  include `half` and `bfloat` in the floating point optimization.
- Add an explicit cast to the type of the `Level` because these types
  don't have "mixed" (e.g taking an `uint` and a `half`) division operators.
  This should not change the behavior for `float` and `double`, because
  this is already what the implicit conversion would do for the division.
- Remove the explicit specialization for `half`, because its unused and
  it can cause problems for consumers of rocprim that have
  `__HIP_NO_HALF_OPERATORS__` defined, in which case the division operator
  on `half` gets removed from the HIP header, and the full specialization
  causes a compilation failure, even when it is not called.

* Add test to verify compiling with `__HIP_NO_HALF_OPERATORS__`

Add a test that verifies a minimal support with
`__HIP_NO_HALF_CONVERSIONS__` and `__HIP_NO_HALF_OPERATORS__`, i.e.
that it compiles with these macros defined.
When these are set operators and conversions on `half` are not defined by
the HIP headers. Some algorithms in rocPRIM will not work with `half`
in this configuration.

[ROCm/rocPRIM commit: d1bbc94cf6d148fdbca0d1fa861a63cd5519177c]