Add ROCSPARSE_MTX_DIR option (#396)
* Use Python from PATH

Using env will look up python3 from the PATH, which allows for the use
of a different version of Python than the one shipped by the OS. That
can be useful, for example, to avoid having to install pyyaml for the
system Python.

* Add ROCSPARSE_MTX_DIR option

When building the rocSPARSE tests using Spack, it's nice to use Spack
to manage the matrix archive files (tar.gz). Spack resources get
cached automatically, so this prevents them from being downloaded
excessively. However, each cached file get stored in a different
subdirectory. Thus, there's no single directory to set as the
ROCSPARSE_TEST_MIRROR. A new option is needed to make this work.

This change adds an option to rocSPARSE to skip the download and
extraction of the matrix archives. Instead, users can specify a
directory where the contents of the archives will be found (the mtx
files). Then rocSPARSE will continue as normal, converting them to csr
for use in the rocSPARSE tests.

Most of the lines touched in this change are whitespace changes due to
changing the indentation level for an extra if. The diff is more clear
with whitespace ignored.

[ROCm/rocSPARSE commit: c2a022d2007418e1f09683d9709c9bd9e061b25e]