Issue Navy32 on the synchroneous split fusion algorithm. (#421)
This algorithm is still in an experimental state, it might be worth to further
 look at it in deep eyes and check if some threadfence could be required.
The fix for now, and I'm very comfortable with it (see below), is to require global 2 iterations minimum.
With the local loop (within a kernel) the synchroneous algorithm becomes asynchroneous in the sense
that the local convergence only involves blocks that are running concurrently, it is dependent of the blocks
scheduling. A second pass does not seem unreasonnable.

[ROCm/rocSPARSE commit: 0884040cda1446b9e3486fc5a64b74e34197bb03]