Fix out-of-bounds access in rocsparse-test (#521)
The rocsparse tests occasionally create a pointer to one past the end
of a vector using &vec[idx] where idx == vec.size(), but this is not a
valid construct. It is valid to create a pointer to one past the end
of a vector, but not by using operator[] to do so. The index argument
must be less than size().

The arguments passed to operator[] can be verified by building with
-D_GLIBCXX_ASSERTIONS.

[ROCm/rocSPARSEcommit: a6e21e6231662aff7f08ad2d00137b07594c84fe]