Use a kernel for small copies to device (#285)
Instead of `hipMemcpyAsync` use a simple kernel to set values in device
memory.
`hipMemcpyAsync` blocks the host until the operations in the stream
complete if the source is not pinned memory.
Using a kernel, the host doesn't have to wait. Note that an atomic write
from the host wouldn't work as the copy has to happen after the current
operations in the stream complete.

[ROCm/rocSPARSE commit: 5db16b39aa3f53e475dd68d271e491579e930122]