Improve error reporting for zero-tolerance test failures (#147)
* Improve error reporting for zero-tolerance test failures

The hipsolver tests report certain error conditions (like unexpected
info values) by adding a positive integer to the measured error. This
can make it difficult to understand what is wrong when reviewing the
logs for a failed set of tests (especially when there are multiple
properties that could all increment the error).

The Google Test EXPECT functions will check a condition, but allow
the test to continue. This means that an error can be recorded at
the place where it occurs, without losing data.

This matches the changes made to rocsolver in cac76478d3a2fb860.

* Fix formatting

* Fix syevd/heevd test

* Fix formatting

[ROCm/hipsolver commit: 560e50c0e91d8e1e9d071d8fa03fe3a6f18c7e56]