Fix stack overflow in sparse tests on Windows (#583)
It appears that the while(p != p_parent) check would create an infinite
loop on Windows, ultimately resulting in a stack overflow. With the
addition of exe-relative lookups, all other searches are superfluous
and can be removed.

* Move sparse data to staging directory

The Windows CI only copies <builddir>/clients/staging to the test
machine, so the data files must be copied to that location.

* Replace std::filesystem::weakly_canonical

On SLES, the fs namespace is an alias for std::experimental::filesystem
and the experimental filesystem implementation does not include
weakly_canonical. Instead, just use the full canonical function,
which will return an error if it does not exist.

(cherry picked from commit 67aea9b005d68c47d7bfb6bc0d1b12a8a34f0fad)


[ROCm/rocSOLVER commit: f0797482fdd0fdf3024d8f9e6f918d9c1de770e2]