Remove librt from MIOpen's public interface (#2287)
librt is not used in miopen's public header so it should only be linked privately and for the tests. 

This causes an issue when using miopen on ubuntu 22.04, because it is named librt.so.1 instead of librt.so. We really shouldn't ever use absolute paths in the target interface since libraries can be located in other locations.

[ROCm/MIOpen commit: 9ad4dc41d519bb74d4e69f251197a9c584c8c042]