build(nvcc): Fix all files being rebuilt when cmake is reconfigured
SetupNVCC was extending `CMAKE_CUDA_FLAGS` and `CMAKE_CXX_FLAGS`
with the same options each time cmake was re-configured which caused
all targets to be re-built every time cmake was configured.

Fix by remembering that it was already done, and not doing it on subsequent
configures. This does mean however that changing these options require
a clean reconfigure, but that was the case previously too.
(The  new options would just get appended leaving the old ones present too)


[ROCm/rocRAND commit: ac4244950bf4c552f4b8700ffc8975e9f5d36c57]