Add rocsolver team as code owners for docs (#668)
When a file matches multiple patterns in the CODEOWNERS file, it is
assigned the code owners listed in the last matching pattern (not
the union of all code owners in all matching patterns). This means that
the additions to CODEOWNERS in 943c6f34ae5b77143f970be628c2e81dc59444ac
implicitly removed the rocsolver team members from documentation files.

[ROCm/rocSOLVER commit: 7477d0b501ba3af009ebdd2a8772e0db868ec01b]