Add hipsolver team as code owners for docs (#229)
When a file matches multiple patterns in the CODEOWNERS file, it is
assigned the code owners listed in the last matching pattern (not
the union of all code owners in all matching patterns). This means that
the additions to CODEOWNERS in 6becf702a618df2e17b287ab501b80c5401f6e61
implicitly removed the hipsolver team members from documentation files.

[ROCm/hipsolver commit: e690e57eeac86f4e984722a90fe33eccaf0aab45]