Add hipsolver team as code owners for docs (#229)
When a file matches multiple patterns in the CODEOWNERS file, it is
assigned the code owners listed in the last matching pattern (not
the union of all code owners in all matching patterns). This means that
the additions to CODEOWNERS in a4f976e794189849c4b289c8f91b25790e3790d1
implicitly removed the hipsolver team members from documentation files.

[ROCm/hipSOLVER commit: e690e57eeac86f4e984722a90fe33eccaf0aab45]