Re-enable hipGraph version of device_adjacent_difference test (#534)
Previously, we disabled the hipGraph version of the LargeIndices test
for the device_adjacent_difference algorithm because it was failing on
Navi3x and MI300.

The root cause of this failure was a call to hipMemset that occasionally
continued to run while the hipGraph was instantiated. I've now learned
that this is expected behaviour - hipMemset runs asynchronously unless the
pointer it is passed refers to pinned host memory. This is true even if
the call to hipMemset is issued into the default stream, like it is in
this case. hipMemset*Async*'s behaviour differs only with respect to
pinned host memory (it's non-blocking in that case). This behaviour is
not currently mentioned in the HIP documentation, so I'll put in a
request to add it there.

This change just inserts a call to hipDeviceSynchronize to ensure that
the hipMemset call completes before the graph is instantiated.
It also removes the GTEST_SKIP() call that disabled the hipGraph version
of the test.

[ROCm/rocPRIM commit: 2dce71a9e57546f2089f45ad8f07147d4398879e]