Use find_library to search for cholmod (#286)
Use find_library to search for cholmod. Newer versions of suitesparse
provide a CHOLMODConfig.cmake, so we can try searching for it first
before adding these new find modules to the CMake module path and
searching with find_library / find_path. The CHOLMOD_ROOT environment
variable can be used as a hint both for config-based searches and
find-module based searches.

[ROCm/hipsolver commit: 643759c740bdfb342e36dd3bb97e8e88d9c5def9]