Add Catch2 tests.
* Add Catch2 tests under test/catch.

It's important to understand the SECTION and GENERATE macros, in particular how the parent sections are repeated for every child section.
Note the TestContext class and the SUPPORTED_ARCH_SECTION macro. I expect we will want a version of SUPPORTED_ARCH_SECTION that will allow us to use this for e.g. all the MFMA archs, or all the gfx12 archs, or something else.
There is a new library common_test_utilities under test/common. The header files are in test/common/common so that you can include them with
#include <common/CommonGraphs.hpp>
The catch2 tests are included in ctest in much the same way as gtest. A couple of differences:
Any parameterizations of the same test will show up as one test to ctest.
There's an open issue where by default skipped tests show up as failed. The SKIP_REGULAR_EXPRESSION arg works around that.
Added a REQUIRE_TEST_TAG assertion that allows checking that the currently running test has a particular tag; this is currently used to check that any test that launches a kernel generated with AssemblyTestKernel has the gpu tag.
Migrated ExpressionTest.cpp to catch2

Split this into 4 files: ExpressionTests.cpp, BinaryExpressionTest.cpp, TernaryExpressionTest.cpp, ScalarExpressionTest.cpp.
We should have more, smaller test files in general.
We may also want to consider splitting up the test executable into multiple smaller ones, particularly for GCC linking time.
There is a new utility class AssemblyTestKernel. See `(Binary|Ternary|Scalar)ExpressionTest.cpp for usage examples. I think this is in general how we should handle tests that create a kernel and run it.
Instead of a function that generates a kernel and launches it once, the class generates the kernel, and then we can launch it different times with different arguments.
Had to figure out some of the quirks of C++ compilers and overflow/casting to get these existing tests to pass with a more complete set of inputs values.
Modified the conditional operator tests to use all of the comparison operators.
Some of these were apparently untested as I had to add those instructions to GPUArchitectureGenerator_defs.hpp.
I'm not sure about decision to make AssemblyTestKernel convert TestContext to a Context.
Changes to rocRoller lib

If you generate an expression with a comment into a register without a name, the register will get the comment as its name. 
Fixed KernelArguments' logging for arguments that are 8-bit integers
They were previously interpreted as c-strings which would usually result in a segfault.
Moved the DefaultArchitectureFilePath generated file from a header in the main repo to a .cpp file in the build directory.
This prevents unnecessary rebuilds and incorrect builds if you have multiple build directories for e.g. clang and gcc.
Added a concatenate_join function that does the same thing as concatenate() but puts a separator in between every argument.