Do not export functions that are defined inline in headers. (#245)
Functions that are defined inline in headers are available to the
including compilation unit. They do not need to be exported from a shared
library for any platform.
Adding a `dllimport` attribute to function definitions is an error on
Windows (at least for MinGW with GCC).

Remove the export (and more importantly the `dllimport`) attributes for
functions that are defined inline in headers to avoid that error.

[ROCm/rocalution commit: 516832c1c6e72d7ed745764ea3d7eddca0a522af]