host_vector.cpp: Define `_USE_MATH_DEFINES` if `_WIN32` is defined. (#250)
This fixes a build issue with MSVC 2022 v17.11:
```
.../rocALUTION/src/base/host/host_vector.cpp:400:62: error: no member named 'M_PI' in 'rocalution::HostVector<float>'
  400 |                             * cos(static_cast<ValueType>(2 * M_PI) * u2);
      |                                                              ^
```

The preprocessor condition for whether to define `_USE_MATH_DEFINES`
currently checks whether `WIN32` is defined. But it is not for that compiler.
It should be enough to check whether `_WIN32` is defined:
http://msdn.microsoft.com/en-us/library/ff540443.aspx

But in other places in the code base more alternative spellings are checked.
So, use the same condition here, too.

[ROCm/rocalution commit: fd5f402a3f02be92d3c16b03ba642f48ea39ca16]