Blockscale Transforms
Added forwarding of scale values from the command through to the assembly, when it's needed.

Currently, many parts are less flexible than we may want them to be in production:

Either both A and B must have scales, or neither.
It won't work for any block sizes other than 32 in the K dimension.
We probably want to do more refactoring to make all 4 inputs (A, A Scale, B, B Scale) follow the same code path instead of copy-pasted blocks. I have done this in some places.

Fix ControlGraph::get<>() and isElemType<>() for Edge/Node.

---------

Co-authored-by: Yiqian Liu <yiqialiu@amd.com>
Co-authored-by: Joao P. L. de Carvalho <joao.carvalho@amd.com>