Enable offload-compress when available (#843)
The size of librocsolver.so.0.4 was reduced from 2,017,868,176 bytes
to 135,433,664 bytes. This is a savings of compression ratio of 14.9
(i.e. a reduction in size of 93%).

Note that the size of the package download is increased by ~25%, as
the compression done by offload compress is less effective than the
compression used for the package file. However, the difference in
download size is minor compared to the benefits in installed size.


[ROCm/rocSOLVER commit: d1be0ce3e47a0a5eb32e9fee7bf3ea50b3907185]