Add IdentifyParallelDimensions graph transformation.
Add IdentifyParallelDimensions graph transformation.

This transformation identifies parallel dimensions:
- Connected to a MatrixMultiply control op
- Connected to LoadTiled/StoreTiled nodes that don't modify the dimensions

Currently, it just reassigns the `size` field of the `SubDimension`s such that they will have the same kernel argument.  We may in the future want to identify this in the graph so that other code can take advantage of this redundancy.  We will also want to add predicates to the CommandKernel so that we check that these sizes are consistent.