Enable Register::Value to represent a bitfield within a register
Bitfields are constructed with either `segment()` or `bitfield()` methods of Value and are lowered to BitFieldExtract expressions. Thus `*_bfe_*` instructions are generated regardless of the fact that a give instruction might support sub-dword addressing (SDWA). SDWA-aware code generation is not implemented yet. Note that creating bitfields with `segment()` is only allowed for packed datatypes. bitfield can be used to create bitfields more generally.