Conditionally log a DeviceGroupedConvBwdWeightTwoStage_Xdl_CShuffle warning (#1860)
The code was emitting a warning if MIOpen did not create a workspace
prior to invoking the IsSupportedArgument method, but the
condition for MIOpen to create a workspace was not met, and so this
condition was not really an error but more of a log message.  This commit
addresses this issue by using the CK_LOGGING facility to only generate the
log message if the CK_LOGGING environment variable is set.