[GFX13] Use __bf16 for bhalf_t when clang version >= 21 (#314)
__bf16 is native supported in clang, and it is widely used in gfx13 built-in functions.
In the original implementation, we use ushort to emulate bhalf, it causes a lot of bit-cast operations in hip source
after this change, the real type of bhalf is changed to __bf16.

known issue: __bf16 isn't well supported on old clang (roc 6.3, clang 19), using __bf16 directly will causes compiler crash when compile some instance, so we only do the switch when clang version >= 21.