Support Swizzle with VectorWidth > 1 (#1720)
This patch supports swizzled tensor with VW > 1 by the following concepts:

Please note that the explanation uses swizzle A as an example.
The notation of A and M should be replaced with B and N when referring to
swizzle B.

For global read A tile offset:
groA0I_0 = WTID * (numSwzBlkK * swzBlkSize) * (VWA * WvM).
*WTID: wave tile ID.
*numSwzBlkK: number of swizzled block in K dimension.
*swzBlkSize: size of a swizzled block.
*WvM: wave group M if swizzle tensor A.

If x and x-1 are in the same VWA group:
groA0I_(x) = groA0I_(x-1) + swzLaneSize.
*swzLaneSize: how many element does a swizzled lane load.

If x and x-1 are in the different VWA group:
groA0I_(x) = groA0I_(x-1) + swzWTOffset.
*swzWTOffset: swizzled wave tile offset.
 Assume X = (numSwzBlockK * swzBlockSize) * (VWA * WvM).
 swzWTOffset = X - swzLaneSize * (VWA - 1).

For global read A unroll offset:
Offset formula (unit: element) = F(r,c) + swzBlkOffsetM(tid)
*r, c: local index of row and column in a swizzled block.
 r = (tid / swzM) * laneSize
 c = (tid % (swzM / VWA)) * VWA
*F(r,c): swizzled offset when vector width A is 1.
 F(r,c) = r % laneSize + c * laneSize + (r / laneSize) * swzM * laneSize
*swzBlockOffsetM(tid) = [tid / (swzMorN / VW)] * swzSizeK
 swzSizeK = numSwzBlkK * numElmInSwzBlk
 numSwzBlkK: number of swizzled block in K dimension.
 numElmInSwzBlk: number of element in a swizzled block.

[ROCm/hipBLASLt commit: 0a7827374297e300d769e9c6e87e5d3d6f44cae5]