Mergeback changes from 6.5 (#223)
* Mergeback 6.4 changes (#219)

* Include 6.2.4 version update (#202)

* Include 6.2.4 version update

* Fix typo

* Cherry-pick to release-staging/rocm-rel-6.4: Update license file for 2025  (#210)

* Update license file for 2025 (#207)

(cherry picked from commit a2c41ca27e830d6c2b5798b5420ef4422394cd69)

* Update license file to include start and end years (#208)

(cherry picked from commit d371ee799695d93528ea877d913b5f6c15f7f619)

* Guard against spaces in HIP_PATH on Windows

If you install the hipSDK on Windows using the installer executable,
the default install location is in "C:\Program Files\...". The installer
also sets the HIP_PATH environment variable to that location.

This can cause problems in CMake, since the space character in the path
causes the path to be interpreted as a list with two items.

This change does the following to fix this problem:

- allow the HIP installation path to be specified using the HIP_PATH
environment variable instead of HIP_DIR (this was changed in
rocRAND a while back, but never made it into hipRAND).

- remove "-I${HIP_DIR}/include" from CMAKE_CXX_FLAGS, since HIP_DIR
can include spaces, which causes the path to be interpreted as
two separate arguments. This include flag is no longer necessary,
and was removed in rocRAND a while back.

- modify rmake.py to bring in updated changes from rocRAND's version
of the script. These changes ensure that the string in the rocm_path
variable is quoted (so spaces won't be an issue). This then causes
CMAKE_PREFIX_PATH to be set with the quoted string and passed into
cmake via the shell call. Passing in the quoted string using the
shell call eliminates the need to quote HIP_PATH / HIP_DIR / CMAKE_PREFIX_PATH
within the hipRAND cmake source files, which I was unable to find
a reliable way to do.

This change also:
- removes some dead commented code in toolchain-windows.cmake.
- increments the hipRAND version number
- adds a note to the changelog to make users aware of the fix

* Update CHANGELOG.md

Co-authored-by: Jeffrey Novotny <jnovotny@amd.com>

* Update CHANGELOG.md

Co-authored-by: Jeffrey Novotny <jnovotny@amd.com>

---------

Co-authored-by: Stanley Tsang <stanley.tsang@amd.com>
Co-authored-by: Jeffrey Novotny <jnovotny@amd.com>
Co-authored-by: ROCm CI Service Account <66695075+rocm-ci@users.noreply.github.com>

* Update version for 6.5, correct changelog entries (#221)

Update the hipRAND version for the ROCm 6.5.0 release.
Move some changelog entries to a 6.4.0 heading (previously
these were mistakenly placed under the 6.2.4 heading.

---------

Co-authored-by: Stanley Tsang <stanley.tsang@amd.com>
Co-authored-by: Jeffrey Novotny <jnovotny@amd.com>
Co-authored-by: ROCm CI Service Account <66695075+rocm-ci@users.noreply.github.com>

[ROCm/hipRAND commit: bc4f9194fab26e162475343cdcb8dff8ce0e32e0]