Reset internal hip error for tests that run out of memory (#90)
The behaviour of hipGetLastError is changing in HIP 7.0. Previously the
error that was reported was cleared on each HIP API call. This means
that hipGetLastError reported any error that occurred during the last
HIP API call.

Moving forward, the error that's reported will only be cleared on each
call to hipGetLastError. This means that hipGetLastError will report any
error that has occurred since the last call to hipGetError.

Some of our tests rely on observing a return value of
hipErrorOutOfMemory from hipMalloc when an allocation is too large for a
given GPU architecture's memory system. This sets the internal HIP
error, and it's not cleared before subsequent tests call
hipGetLastError, causing them to fail.

This change adds extra calls to hipGetLastError to clear the error (for
future tests) in cases where tests run out of memory.