[rocPRIM] Avoid creating void reference type in device batch memcpy (#135)
In the batch_memcpy_impl struct, we define two types based on the value
of the InputBufferItType template argument.
1. input_type, set to the InputBufferItType's underlying value_type's
value_type.
2. Alias, a type that's set using a std::conditional statement that
examines the IsMemCpy boolean template arg, and may be set to either
unsigned char or the type from (1).

Later code creates a reference to (1). This causes problems when (1) is
void.

This change removes the definition of (1), since it does not appear to
be used anywhere else in our repos (it's part of the detail namespace,
so it's not public).

However, this is not enough to fix the problem, since the compiler
evaluates both sides of the std::conditional we use to define (2). To
work around this, this change also adds a helper struct type,
(AliasType) that uses template specialization on the IsMemCpy boolean to
define the type that Alias will be assigned. This allows us to remove
the std::conditional statement.