fix(rocprim): fix incorrect masking in dpp warp scan (#78)
This change resolves a test failure observed in a future rocThrust test.
This is a blocker for adding CCCL 2.7 parity in rocThrust and should be
included in ROCm 7.0.

It adds 2 changes:

- Skip including rocprim_version.hpp when `_CLANGD` is defined. This
resolves `clangd` complaining about things.
- Fix incorrect masking in DPP warp scan.

---

Originally, executing the scan operator was not masked. That means that
all lanes could potentially execute. In the case of the failing thrust
test, it involved some pointer chasing. My guess is that some
optimization logic would still execute the scan operation (read: pointer
chasing) even though the input of the operation is optimized out. Since
we don't know what the values of those registers are when we do the scan
operation, it may chase the pointer to something in the middle of
nowhere, causing a crash.