[rocThrust] Add fallback to std::decay_t when building against old libstdc++ (#200)
The recent updates for CCCL 2.7 compatibility introduced a call to the
built-in function std::__decay_t. This built-in can be used in place of
std::decay_t to reduce compilation time. It is defined starting with
libstdc++-10. However, RHEL 8 still uses libstdc++-8.x.

This change introduces a fall-back to std::decay_t in cases where we're
building against an older libstdc++.