Reverse win gtest condition (#450) (#505)
Reverse UseGTestAssert condition on Windows
    
Recently, we added a check to see if GTest's ASSERT_EQ assertion should
be used within the assert_eq function. The code in the if/else blocks
that act on the results of this check was inverted (the "else" code
block should be the "if" block, and vice-versa). This change fixes the
issue by swapping the code blocks.

This change should be merged after #449.