Cherry-pick to release/rocm-rel-7.0: [hiprand] Docs: Update repository and path information for rocm-libra… (#642)
…ries switchover (#233)

This is a companion to https://github.com/ROCm/rocm-libraries/pull/232.
It updates the hiprand docs index page, install instructions, and links
to the GitHub to reference rocm-libraries rather than the legacy hipRAND
repo. It also adds a note directing users of pre-7.0 branches to the old
repository and documentation.

NOTE: This is intended for the 7.0 branch and later. A different change
will be applied to the 6.4 branch explaining the situation from the
perspective of older branches.

(cherry picked from commit f588fa73bb96a7a2b6ff384822e8d53cc90a8462)