Temporary workaround for compiler issue on Navi3x (#701)
This is a temporary workaround for Navi3x while we wait for a compiler
fix. This issue appears to cause zeros in data loaded using the
block_load_direct_striped method. Changing way the data is accessed, or
removing ROCPRIM_UNROLL seems to work around it. This change takes the
former approach.