[rocsparse] Fix Windows compilation and linking failures (#794)
Summary of proposed changes:
- This PR fixes Windows compilation and linking failures.
- The compilation errors were related to the compiler not knowing which
`std::fma` to use because `rocsparse_bfloat16` could be implicitly
casted to `float` and `double`. The solution was to only allow implicit
casting `rocsparse_bfloat16` to `float` (which seems reasonable) but not
to `double`. Casting to `double` now requires a `static_cast`.
- There were also ODR linking failures introduced in this PR:
https://github.com/ROCm/rocSPARSE/pull/535/files. I have instead made
the templated function pointers inline. This change still maintains the
original PR's purpose of getting rid of warnings.