[hipBLAS] Add Clang guards around usage of openmp config (#1050) (#832)
This fixes a gcc issue where clang headers are being brought in by the
openmp-config.

Additional cleanup:
Remove unnecessary options added to COMMON_LINK_LIBS. If the openmp
config is not found then OpenMP::OpenMP_CXX will be the fallback for
both gcc and clang.

[ROCm/hipBLAS commit: 46391031ec86a8edc7874d7dc8f35c4923134f05]

This patch is dependent on the Ubuntu 24.04 docker file being updated to
use AOCL 4.2 instead of 4.0.

Co-authored-by: Torre Zuk <42548444+TorreZuk@users.noreply.github.com>