[rocsparse] make sure to find a correct version of rocprim and to use the correct header files even if another version is installed (#1074)
When compiling rocSPARSE on machines that have rocm preinstalled, users
can run into issues where a preinstalled version of rocprim is used
during the configuration/compilation.
The preinstalled version of rocprim is not necessarily consistent with
the version included in the repo and used by rocSPARSE.
On top of that, even if a correct version is found, it happens that
during the compilation phase the header files of the older version
(installed in /opt/) are used.
This PR resolves both issues.