[rocprim] Optimize device_select, device_partition for MI3xx multi-stream performance (#1044)
Credit goes to @Naraenda and @NB4444 for writing the original fix for
6.4.

Under certain conditions, device_select and device_partition can
experience slowdowns on MI3xx GPUs when running on multiple streams.
This fix mitigates the slowdown by utilizing atomic counters instead of
flat block id's to assign work. To preserve performance when the user
knows in advance that multiple streams are not being used, or on non
MI3xx-based architectures, this optimization is opt-in only, via
template argument.