[rocPRIM][Code Coverage][Cherry Pick] Increase code coverage for rocPRIM (#1036)
In this commit we have added test coverage, made some bug fixes and
added a wrapper function for a common pattern in the tests. This
`test_kernel_wrapper` function reduces the amount of code used.

In

https://github.com/ROCm/rocm-libraries/pull/242/commits/0318f427d0eab1f29c331a6c16a0371e3e4f1724
we have added coverage of the `warp_scan_shuffle` by forcing dpp off in
cmake on the `warp_scan` test.

In

https://github.com/ROCm/rocm-libraries/pull/242/commits/30d54c18bb662c7e082c62f30ded9f17a84cd1bd
we have added the missing test cases for the thread_algos in the thread
directory. (Except for thread_operators, these functions are mostly
duplicates probably needs to be cleanly moved/deprecated). We also found
a bug in `thread_reduce` with the tests and fixed this.

In

https://github.com/ROCm/rocm-libraries/pull/242/commits/48780ff57449c9315bf4c4a42e8c338aaa7e4b78
https://github.com/ROCm/rocm-libraries/pull/242/commits/37aa542fb924bc4f76310d589b56fe94231b199b
https://github.com/ROCm/rocm-libraries/pull/242/commits/05a9e12013a5ab9c44e6b9f5a5a7451156e8cade
https://github.com/ROCm/rocm-libraries/pull/242/commits/f5f1dbe7ac5944d508c6f5207e7279607dd1a71a
https://github.com/ROCm/rocm-libraries/pull/242/commits/343e275b619ee94032590ec3a760aaaae85f8c1f
https://github.com/ROCm/rocm-libraries/pull/242/commits/8007fe6c466e90fd007df3926c3692a9fadc10a1
we have added unit tests for the files in the types directory.

In

https://github.com/ROCm/rocm-libraries/pull/242/commits/40ca3e2cab44b8d59fd0fef2ad6988de3a5dbda2
we have introduced the `test_kernel_wrapper` and added the missing
`PartitionTwoWayFlag` test.

In

https://github.com/ROCm/rocm-libraries/pull/242/commits/7a4e7ba4ccbb7f45aa10513a4b3ac5285936bf62
unit tests for the rocprim::tuple type are added.

In

https://github.com/ROCm/rocm-libraries/pull/242/commits/29eac9abbb08c64cbeb6ee344e1a1ae69996f2fb
we have added extra test coverage by introducing a new type that will go
into the untested path. We have also added this type to the benchmark to
show [this

specialization](https://github.com/ROCm/rocm-libraries/pull/242/commits/29eac9abbb08c64cbeb6ee344e1a1ae69996f2fb#diff-22a70b2ad081732e222004ded43ce0db7145ff196f7b723289425dcb5b6c732dR228)
is still needed. We also changed the `std::is_integral` to
`rocprim::is_integral` to not include `(u)int128_t` for this
specialization, this does not impact performance. The specialization
enable_if was also slightly changed to make it clearer which path is
chosen (does not make a difference in actual executed code). Also custom
config and a iterator was added to the tests of merge_sort.

In

https://github.com/ROCm/rocm-libraries/pull/242/commits/f5614d585d521fbcee8ae324dcda783f1181253e
test coverage was added for the device_scan_common.hpp file.

In

https://github.com/ROCm/rocm-libraries/pull/242/commits/1d2a40aaa6df6b7a4dc6bdf5acb2b1054fbc9b92
test coverage was increased by actually using `const fixed_array` and
including a `Level` type which goes into the base path of
sample_to_bin_even struct. We also added an iterator as a type to the
test. Also the new test wrapper was used.

In

https://github.com/ROCm/rocm-libraries/pull/242/commits/c518f4259bac8541232d5758efe2bc46b68a87c4
test coverage was increased for the iterators, a lot of the operators
where missing. Also some cleaning up of the tests was done including the
wrapper when possible. There was also a bug found for the comperator in
`arg_index_iterator` and `texture_cache_iterator`, which was also fixed.
The `->` operator is not tested for `test_texture_cache`,
`arg_index_iterator`, `transform_iterator` and `zip_iterator`. They
currently do not seem to work, I am working on a possible fix but will
be added in a later PR.

---------

Reopening this PR to trigger all CIs and labels. Original PR here:
(#753)

---------

Co-authored-by: Nick Breed <78807921+NB4444@users.noreply.github.com>
Co-authored-by: Saiyang Zhang <saiyang@streamhpc.com>
Co-authored-by: Nara Prasetya <nara@streamhpc.com>
Co-authored-by: Cenxuan Tian <cenxuan@streamhpc.com>