[TheRock CI] Adding docker image for TheRock tests (#1307)
Previously, these tests were running on machines with an already
installed ROCm.

Now, this docker image on Linux ensure that ROCm is not installed and we
are getting ROCm from TheRock

Working here:
https://github.com/ROCm/rocm-libraries/actions/runs/17135064877

(this will be testable in PR after #1306 goes through. Also, I plan to
add documentation to TheRock on how to replicate errors with this docker
image)