[hipRAND] Update CUDA compute capability detection with nvidia-smi based fallback (#1286)
## Motivation

For some reason, the current method of detecting CC in SetupNVCC.cmake
has weird, unexplained interactions with the math-ci environment. I've
decided to add a fallback which uses nvidia-smi to detect the CC(s) of
the current GPU(s) installed on the machine, which should not be subject
to weird interactions. It'll take the minimum CC if multiple are
detected.

## Technical Details

The new fallback method does require CUDA Toolkit 11.6 or newer to work,
but fortunately that is not a problem since math-ci uses CUDA 12. And,
it is a fallback method so it does not break existing workflow.

## Test Plan

Verify math-ci works.

## Test Result

It works.

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.