[superbuild] Add pipeline trigger (#1375)
## Motivation

In order to have testing on the first superbuild PR #1239 this trigger
file is required.

## Technical Details

Adds branch and PR triggers only for the top-level CMakeLists.txt and
cmake directory.

## Test Plan

Testing for this trigger will be conducted in #1239, which will also
update the relevant trigger directories for the supported projects.

## Test Result

Passed