Improved stedc (reshuffle gemm) (#1017)
Code refactor and improvements to gemm calls.

- Original code used monolithic gemms that cover the entire matrix.
Breaking them down to only work on blocks improves performance.
- Eigenvalues are now sorted. They are sorted before deflation to make
deflation code faster, and after deflations to group non deflated values
together. That allows to simplify and speedup the rest of the code.
- There are redundant data movements of D/ev and vectors (the reason is
to localize the changes and structurally mimic original processing to
reduce risks of a big refactoring). Optimizing it should provide
additional (but probably minor) performance boost and simplify code. But
it is best to postpone it to a separate PR
- Performance of the secular eq solver was improved.
- Performance of the mergeVectors was significantly improved.

---------

Co-authored-by: Ilya Perminov <Ilya.Perminov@amd.com>
(cherry picked from commit 0832dbf5fabc04633cb9597faf38f1bdaf047395)


[ROCm/rocSOLVER commit: 5b0116806c480df46ddcdd7fd0563fe923faa4a4]