[hipCUB] Pull rocPRIM dependency from monorepo, add new fetch option (#1123)
hipCUB has migrated to the new monorepo repository. This change:
1. Updates cmake files to cause the configuration process to download
rocPRIM from the new monorepo instead of the old public repo.
2. Adds new cmake options so that, if the user is building from a
monorepo checkout (which already provides rocPRIM), they can just use
the source code in the monorepo instead of downloading it.
3. Adds the ability to set the rocPRIM branch that should be downloaded.
4. Updates documentation to reflect the above.

For (2), the new cmake build option is ROCPRIM_FETCH_METHOD. This option
may be set to one of three values:
- PACKAGE (default) - search for the dependency in an installed package.
If no package with the minimum required version is found, the method
falls back to DOWNLOAD (below).
- DOWNLOAD - download the dependency from the monorepo. Since the
monorepo is large, if it is possible, we don't want to download the
whole thing. If git >= 2.25 is present, this option will use git
sparse-checkout to significantly reduce the size of the download (only
the project/<repo_name> directory will be fetched). If we're limited to
an older version of git, a warning will be displayed and the whole
monorepo will be downloaded. In both cases, on subsequent builds we
check the build directory for an existing copy of the dependency to
avoid downloading again if possible.
- MONOREPO - indicates that we are building from a monorepo checkout.
Searches for the dependency in the projects folder at
../../projects/<repo_name>. If the dependency is not found, cmake will
check if this is a sparse-checkout. If so, it will try to add rocprim to
the sparse-checkout list. If something goes wrong, this option falls
back to DOWNLOAD (above).

This change also renames the existing DEPENDENCIES_FORCE_DOWNLOAD build
option to EXTERNAL_DEPS_FORCE_DOWNLOAD in order to differentiate it from
the new rocPRIM dependency option mentioned above. The behaviour of this
option is also changed slightly - when enabled, it no longer forces a
download of rocPRIM, only GTest and GBench.

---------

Co-authored-by: Stanley Tsang <stanley.tsang@amd.com>