Replace gfx942 configs with gfx942 variant configs
We currently have no way of distinguishing between different variants of gfx942.
This means that the gfx942 configs are used on all gfx942 family architectures.
In some cases this is not optimal. This change replaces the existing configs
with specialized ones for a specific variant, as per a customer request.

This change should not be merged to develop or release branches - it is provided
solely for customer use. We'll follow up soon with some config-system refactoring
that allows us to properly distinguish between gfx942 variants so that we can
supply separate configs for each.