[TheRock CI] Honoring workflow dispatch arguments (#1828)
Currently for pull_request, workflow_dispatch and is_push, it runs all
projects if a `.github/workflows/therock-*` or
`.github/scripts/therock-*` is changed.

However for workflow_dispatch, we don't want to run all tests
(especially when adding a new test for CI) during testing as it takes up
a lot of machine utilization.

Works here for workflow_dispatch run:
https://github.com/ROCm/rocm-libraries/actions/runs/18020376566