rocfft: include batch dimension in MPI sample bricks (#1785)
## Motivation

Fix the rocFFT MPI example.  Currently it fails to create the FFT plan.

## Technical Details

An N-dimensional distributed FFT is specified to rocFFT as a set of
bricks. However, the bricks require N+1 indexes for strides and brick
coordinates. The bricks were incorrectly specified with N indexes.

## Test Plan

Ran the example manually with the fix. MPI unit tests cover additional
cases in the correct (N+1) way.

## Test Result

Manual run was successful.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.