[MIOpen] Remove reboot from all MIOpen CI stages (#1984)
## Motivation

Some gfx90a stages were still causing reboots on CI machines because
they weren't supply a value for the needs_reboot parameter and it was
defaulting to true. We no longer need to reboot machines during CI runs
and it was causing failures in other CI runs that were running
concurrently.

## Technical Details

Removed the reboot function and all references to it, as well as the
needs_reboot parameter.

## Test Plan

Doing a standard CI run and checking for reboots

## Test Result

MIOpen CI run successful, no reboots

## Submission Checklist

- [X] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.